refactor: extract script templates to src/tools/helpers/templates.ts
